Update ContactPropertyVersion.sql
Displays all web activity, including profile updates, that happens to a Contact. This includes what pages they look at, when a value of their profile updates, and more. They must be cookied and have a contact profile with us in Hubspot for us to see this information. This is a relatively large table.