Create Topic Modeling.py
This is a LDA Topic Modeling based script that was used to analyze high-level topics of the Contact Us Form Submissions. This is a preliminary analysis that should not be used as a standalone code & deliverable. Note, the text file that this code corresponds to is not attached, and thus the code will need to be modified slightly if you're using this for any other data set.